Restructuring to match other OpenET modules
Switched module name from geerefet to refetgee
