+++ README.rst
@@ -1,10 +1,10 @@
-=======================================================================
-Google Earth Engine ASCE Standardized Reference Evapotranspiration (ET)
-=======================================================================
+================================================================================
+OpenET - Google Earth Engine ASCE Standardized Reference Evapotranspiration (ET)
+================================================================================
 
 |version| |build|
 
-Google Earth Engine (GEE) functions for computing daily and hourly reference ET following the ASCE Standardized Reference Evapotranspiration Equations (ASCE2005_).
+This repository provides `Google Earth Engine <https://earthengine.google.com/>`__ Python API based implementation of the ASCE Standardized Reference Evapotranspiration Equations (ASCE2005_) for computing daily and hourly reference ET.
 
 Usage
 =====
@@ -20,7 +20,7 @@ The raw input data is available `here <https://www.usbr.gov/pn-bin/daily.pl?stat
 
     import math
     import ee
-    import geerefet
+    import refetgee
 
     # Unit conversions
     tmin_c = (66.65 - 32) * (5.0 / 9)                          # F -> C
@@ -31,7 +31,7 @@ The raw input data is available `here <https://www.usbr.gov/pn-bin/daily.pl?stat
     uz = 4.80 * 0.44704                                        # mpg -> m s-1
     lat = 39.4575                                              # degrees
 
-    etr = geerefet.Daily(
+    etr = refetgee.Daily(
         tmin=tmin_c, tmax=tmax_c, ea=ea, rs=rs, uz=uz, zw=3, elev=1208.5,
         lat=lat, doy=182).etr().getInfo()
 
@@ -48,7 +48,7 @@ The raw input data is available `here <https://www.usbr.gov/pn-bin/instant.pl?st
 
     import math
     import ee
-    import geerefet
+    import refetgee
 
     # Unit conversions
     tmean_c = (91.80 - 32) * (5.0 / 9)           # F -> C
@@ -58,7 +58,7 @@ The raw input data is available `here <https://www.usbr.gov/pn-bin/instant.pl?st
     lat = 39.4575                                # degrees
     lon = -118.77388                             # degrees
 
-    etr = geerefet.Hourly(
+    etr = refetgee.Hourly(
         tmean=tmean_c, ea=ea, rs=rs, uz=uz, zw=3, elev=1208.5,
         lat=lat, lon=lon, doy=182, time=18).etr().getInfo()
 
@@ -72,10 +72,10 @@ A helper function for computing daily ETo and ETr for `GRIDMET <http://www.clima
 .. code-block:: console
 
     import ee
-    import geerefet
+    import refetgee
 
     gridmet_img = ee.Image(ee.ImageCollection('IDAHO_EPSCOR/GRIDMET').first())
-    etr = geerefet.Daily.gridmet(gridmet_img).etr()\
+    etr = refetgee.Daily.gridmet(gridmet_img).etr()\
         .reduceRegion(reducer=ee.Reducer.first(),
                       geometry=ee.Geometry.Point(-118.77388, 39.4575),
                       scale=1000)\
@@ -93,11 +93,11 @@ For the daily function, the NLDAS collection must be filtered to a single 24 hou
 .. code-block:: console
 
     import ee
-    import geerefet
+    import refetgee
 
     nldas_coll = ee.ImageCollection('NASA/NLDAS/FORA0125_H002')\
         .filterDate('2015-07-01', '2015-07-02')
-    etr = geerefet.Daily.nldas(nldas_coll).etr()\
+    etr = refetgee.Daily.nldas(nldas_coll).etr()\
         .reduceRegion(reducer=ee.Reducer.first(),
                       geometry=ee.Geometry.Point(-118.77388, 39.4575),
                       scale=1000)\
@@ -108,10 +108,10 @@ For the daily function, the NLDAS collection must be filtered to a single 24 hou
 .. code-block:: console
 
     import ee
-    import geerefet
+    import refetgee
 
     nldas_img = ee.Image(ee.ImageCollection('NASA/NLDAS/FORA0125_H002').first())
-    etr = geerefet.Hourly.nldas(nldas_img).etr()\
+    etr = refetgee.Hourly.nldas(nldas_img).etr()\
         .reduceRegion(reducer=ee.Reducer.first(),
                       geometry=ee.Geometry.Point(-118.77388, 39.4575),
                       scale=1000)\
@@ -203,11 +203,20 @@ The main difference between the two "methods" is that the "asce" method attempts
 Installation
 ============
 
-To install the RefET-GEE python module:
+The OpenET RefET GEE python module can be installed via pip:
 
 .. code-block:: console
 
-    pip install geerefet
+    pip install openet-refet-gee
+
+OpenET Namespace Package
+========================
+
+Each OpenET model is stored in the "openet" folder (namespace).  The model can then be imported as a "dot" submodule of the main openet module.
+
+.. code-block:: console
+
+    import openet.refetgee as refetgee
 
 Validation
 ==========
@@ -232,9 +241,9 @@ References
  | ASCE-EWRI (2005). The ASCE standardized reference evapotranspiration equation.
  | `https://ascelibrary.org/doi/book/10.1061/9780784408056 <https://ascelibrary.org/doi/book/10.1061/9780784408056>`__
 
-.. |build| image:: https://github.com/cgmorton/RefET-GEE/workflows/build/badge.svg
+.. |build| image:: https://github.com/Open-ET/openet-refet-gee/workflows/build/badge.svg
    :alt: Build status
-   :target: https://github.com/cgmorton/RefET-GEE
-.. |version| image:: https://badge.fury.io/py/geerefet.svg
+   :target: https://github.com/Open-ET/openet-refet-gee
+.. |version| image:: https://badge.fury.io/py/refetgee.svg
    :alt: Latest version on PyPI
-   :target: https://badge.fury.io/py/geerefet
+   :target: https://badge.fury.io/py/refetgee
