+++ openet/refetgee/__init__.py
@@ -1,4 +1,4 @@
 from .daily import Daily
 from .hourly import Hourly
 
-__version__ = "0.1.14"
+__version__ = '0.1.15'
