+++ openet/refetgee/daily.py
