+++ requirements_test.txt
@@ -1,3 +1,6 @@
 pytest>=4.0.0
+pytest-cov
+coverage
+codecov
 pandas
 pytz
